src/main/java/com/iGame/AuthorProperties.java
package com.iGame;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;

@Component
public class AuthorProperties {
	
	@Value("${author}")
	private String author;
	@Value("${project}")
	private String project;
	
	public AuthorProperties(){
	}
	
	public String getAuthor() {
		return author;
	}
	public void setAuthor(String author) {
		this.author = author;
	}
	public String getProject() {
		return project;
	}
	public void setProject(String project) {
		this.project = project;
	}

}
